add node to start

+++ 0x12-singly_linked_lists/2-add_node.c
@@ -7,7 +7,7 @@
 */
 list_t *add_node(list_t **head, const char *str)
 {
-	int len_str = 0;
+	size_t len_str = 0;
 	list_t *new;
 	char *dup_str;
